compile time dependency injection

+++ app/uk/gov/hmrc/gform/models/helpers/DateHelperFunctions.scala
@@ -50,7 +50,7 @@ object DateHelperFunctions {
 
       val dateWithOffset = LocalDate
         .parse(dateExprToString(dateExpr), dateFormatter)
-        .plusDays(nonZero)
+        .plusDays(nonZero.toLong)
 
       convertToDateExpr(dateWithOffset)
   }
